Combine apis into index.ts
