+++ prisma/schema.prisma
@@ -22,6 +22,7 @@ model User {
   password   String
   email      String      @unique
   Admin      Admin?
+   type       UserType
   NormalUser NormalUser?
   Superadmin Superadmin?
   Post Post[]
@@ -46,6 +47,12 @@ model Superadmin {
   FK               User?   @relation(fields: [userPK], references: [PK])
   permission_level Int?    @default(2)
   userPK           String? @unique
+}
+
+enum UserType {
+  normalUser
+  admin
+  superAdmin
 }
 
 model Post {
